refactor: scheduleTypeOptions 파일 constants 폴더로 이동 #76
