+++ front/app/constants/scheduleTypeOptions.tsx
